Refactor and clean up the workspace
