+++ Cargo.toml
@@ -1,10 +1,9 @@
 [workspace]
-members = [
-  "twine-core",
-  "twine-macros",
-  "twine-components",
-  "twine-engine",
-  "integration-tests",
-]
-
 resolver = "2"
+
+members = [
+  "twine-components",
+  "twine-core",
+  "twine-examples",
+  "twine-macros",
+]
