--- integration-tests/Cargo.toml
@@ -1,14 +1,0 @@
-[package]
-name = "integration-tests"
-version = "0.1.0"
-edition = "2021"
-publish = false
-
-[dependencies]
-twine-core = { path = "../twine-core" }
-twine-components = { path = "../twine-components" }
-twine-macros = { path = "../twine-macros" }
-serde = { version = "1.0", features = ["derive"] }
-serde_json = "1.0.136"
-serde_yaml = "0.9.34"
-toml = "0.8.19"
