+++ twine-components/Cargo.toml
@@ -1,13 +1,13 @@
 [package]
 name = "twine-components"
-version = "0.1.0"
+version = "0.1.1"
 edition = "2021"
 license = "MIT"
 authors = ["Isentropic Development <info@isentropic.dev>"]
 description = "A collection of components for Twine, a Rust framework for functional and composable system modeling."
 repository = "https://github.com/isentropic-dev/twine"
 readme = "../README.md"
-keywords = ["twine", "components", "functional", "composable", "modeling"]
+keywords = ["twine", "framework", "functional", "composable", "modeling", "components"]
 
 [dependencies]
 twine-core = { version = "0.1", path = "../twine-core" }
